backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "uvicorn>=0.41.0",
]
